+++ keyboards/crkbd/keymaps/jrkopp/include/jk_keycodes.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "keycodes.h"
+
+// US ANSI shifted keycode aliases
+#include "keymap_us.h"
+
+#include "quantum_keycodes.h"
+
+enum custom_keycodes {
+    KC_LT_LP = SAFE_RANGE,
+    KC_GT_RP,
+    KC_AM_EX,
+    MNW,
+    MSW,
+    MSE,
+    MNE
+};
