+++ keyboards/crkbd/keymaps/jrkopp/include/jk_layers.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum layer_names {
+    BASE = 0,
+    SYMBOL,
+    NAV,
+    MOUSE,
+    MEDIA,
+};
+
