+++ keyboards/crkbd/keymaps/jrkopp/include/jk_macros.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#define LGA LGUI_T(KC_A)
+#define LAS LALT_T(KC_S)
+#define LCD LCTL_T(KC_D)
+#define LSF LSFT_T(KC_F)  
+#define RGSC RGUI_T(KC_SCLN)
+#define RAL RALT_T(KC_L)
+#define RCK RCTL_T(KC_K)
+#define RSJ RSFT_T(KC_J)
+#define CBS LCTL(KC_BSPC)
+#define CRA LCTL(KC_RIGHT)
+#define CLA LCTL(KC_LEFT)
+#define CDEL LCTL(KC_DEL)
+#define CHOME LCTL(KC_HOME)
+#define CEND LCTL(KC_END)
+#define UNDO LCTL(KC_Z)
+#define COPY LCTL(KC_C)
+#define CUT LCTL(KC_X)
+#define PASTE LCTL(KC_V)
