+++ keyboards/crkbd/rules.mk
@@ -3,4 +3,6 @@ DEFAULT_FOLDER = crkbd/rev1
 COMBO_ENABLE = yes
 CAPS_WORD_ENABLE = yes
 
-SRC += features/achordion.c
+SRC += custom/custom_achordion.c features/achordion.c
+
+CFLAGS += -Ikeyboards/crkbd/keymaps/jrkopp/include -Ikeyboards/crkbd/keymaps/jrkopp/features
